refactor: move base definitions of uconvert, ustrip
