+++ src/unxt/_src/quantity/functional.py
@@ -1,52 +1,132 @@
 """Functional approach to Quantities."""
 
-__all__ = ["uconvert", "ustrip"]
+__all__: list[str] = []
 
-from typing import Any
 
+from jaxtyping import Array
 from plum import dispatch
 
+from dataclassish import replace
 
-@dispatch.abstract  # type: ignore[misc]
-def uconvert(u: Any, x: Any, /) -> Any:
+from .base import AbstractQuantity
+from unxt._src.dimensions.core import dimensions_of
+from unxt._src.units.core import AbstractUnits, units
+from unxt._src.units.system.base import AbstractUnitSystem
+
+# ===================================================================
+# Convert units
+
+
+@dispatch
+def uconvert(u: AbstractUnits, x: AbstractQuantity, /) -> AbstractQuantity:
     """Convert the quantity to the specified units.
 
     Examples
     --------
-    >>> from unxt import Quantity, uconvert, units
+    >>> from unxt import Quantity, units
 
-    >>> q = Quantity(1, "km")
-    >>> uconvert(units("m"), q)
-    Quantity['length'](Array(1000., dtype=float32, ...), unit='m')
-
-    >>> uconvert("m", q)
-    Quantity['length'](Array(1000., dtype=float32, ...), unit='m')
-
-    For further examples, see the other method dispatches.
+    >>> x = Quantity(1000, "m")
+    >>> uconvert(units("km"), x)
+    Quantity['length'](Array(1., dtype=float32, ...), unit='km')
 
     """
-    raise NotImplementedError  # pragma: no cover
+    # Hot-path: if no unit conversion is necessary
+    if x.unit == u:
+        return x
+
+    # TODO: jaxpr units so we can understand them at trace time.
+    # Hot-path: if in tracing mode
+    # if isinstance(x.value, jax.core.Tracer) and not can_convert_unit(x.unit, u):
+    #     return x.value
+
+    # Astropy correction factor
+    # TODO: this only works with multiplicative unit conversions
+    factor = x.unit.to(u)
+
+    return replace(x, value=x.value * factor, unit=u)
 
 
-# ===================================================================
-
-
-@dispatch.abstract  # type: ignore[misc]
-def ustrip(u: Any, x: Any, /) -> Any:
-    """Strip the units from the quantity, first converting if necessary.
+@dispatch
+def uconvert(ustr: str, x: AbstractQuantity, /) -> AbstractQuantity:
+    """Convert the quantity to the specified units.
 
     Examples
     --------
-    >>> from unxt import Quantity, uconvert, units
+    >>> from unxt import Quantity, units
 
-    >>> q = Quantity(1, "km")
-    >>> ustrip(units("m"), q)
-    Array(1000., dtype=float32, ...)
-
-    >>> ustrip("m", q)
-    Array(1000., dtype=float32, ...)
-
-    For further examples, see the other method dispatches.
+    >>> x = Quantity(1000, "m")
+    >>> uconvert("km", x)
+    Quantity['length'](Array(1., dtype=float32, ...), unit='km')
 
     """
-    raise NotImplementedError  # pragma: no cover
+    return uconvert(units(ustr), x)
+
+
+@dispatch
+def uconvert(usys: AbstractUnitSystem, x: AbstractQuantity, /) -> AbstractQuantity:
+    """Strip the units from the quantity.
+
+    Examples
+    --------
+    >>> from unxt import Quantity, units
+    >>> from unxt.unitsystems import galactic
+
+    >>> q = Quantity(1e17, "km")
+    >>> uconvert(galactic, q)
+    Quantity['length'](Array(3.2407792, dtype=float32, ...), unit='kpc')
+
+    """
+    return uconvert(usys[dimensions_of(x)], x)
+
+
+# ===================================================================
+# Strip units
+
+
+@dispatch
+def ustrip(u: AbstractUnits, x: AbstractQuantity, /) -> Array:
+    """Strip the units from the quantity.
+
+    Examples
+    --------
+    >>> from unxt import Quantity, units
+
+    >>> q = Quantity(1000, "m")
+    >>> ustrip(units("km"), q)
+    Array(1., dtype=float32, ...)
+
+    """
+    return uconvert(u, x).value
+
+
+@dispatch
+def ustrip(u: str, x: AbstractQuantity, /) -> Array:
+    """Strip the units from the quantity.
+
+    Examples
+    --------
+    >>> from unxt import Quantity
+
+    >>> q = Quantity(1000, "m")
+    >>> ustrip("km", q)
+    Array(1., dtype=float32, ...)
+
+    """
+    return uconvert(units(u), x).value
+
+
+@dispatch
+def ustrip(u: AbstractUnitSystem, x: AbstractQuantity, /) -> Array:
+    """Strip the units from the quantity.
+
+    Examples
+    --------
+    >>> from unxt import Quantity
+    >>> from unxt.unitsystems import galactic
+
+    >>> q = Quantity(1e17, "km")
+    >>> ustrip(galactic, q)
+    Array(3.2407792, dtype=float32, weak_type=True)
+
+    """
+    return ustrip(u[dimensions_of(x)], x)
