+++ src/unxt/_src/quantity/register_dispatches.py
@@ -9,9 +9,9 @@ from jaxtyping import ArrayLike
 from plum import dispatch
 from plum.parametric import type_unparametrized as type_np
 
+from .api import ustrip
 from .base import AbstractQuantity
 from .core import Quantity
-from .functional import ustrip
 
 # -----------------------------------------------
